refactor(story): convert /story/[slug] to TypeScript, split data layer into modules/story
Ports the story page to .tsx per the README's TS/module architecture:
route composition stays in pages/story/[slug].tsx, GSSP data fetching
moves to modules/story/story-data.ts, and the client-side "load more
related stories" behavior moves to its own hook backed by a
client-safe fetch module (kept separate from story-data.ts so its
Node-only fs dependency doesn't leak into the browser bundle).

Also fills in the MisoProduct typedef's missing title/url/cover_image
fields and adds an ambient module declaration for the untyped
lilith-draft-renderer import.
